+++ _unittests/ut_onnxrt/test_onnxrt_simple_voting_regressor.py
@@ -37,8 +37,7 @@ class TestOnnxrtSimpleVotingRegressor(ExtTestCase):
         X_test = numpy.vstack([X_test[:4], X_test[-4:]])
         res0 = clr.predict(X_test).astype(numpy.float32)
 
-        model_def = to_onnx(clr, X_train.astype(numpy.float32),
-                            dtype=numpy.float32)
+        model_def = to_onnx(clr, X_train.astype(numpy.float32))
 
         oinf = OnnxInference(model_def, runtime='python')
         res1 = oinf.run({'X': X_test})
